6502 is now 32 bits

+++ target-6502/helper.h
@@ -1,11 +1,11 @@
 #include "def-helper.h"
 
-DEF_HELPER_1(printchar, void, i64)
+DEF_HELPER_1(printchar, void, i32)
 DEF_HELPER_0(getnum, tl)
 
 
 
-DEF_HELPER_2(printstuff, void, i64, i32)
+DEF_HELPER_2(printstuff, void, i32, i32)
 
 DEF_HELPER_2(excp, void, int, int)
